refactor(utils): move keyboard_evts.py to mana_rich_text_ctrl
Only consumer is widgets/panels/mana_rich_text_ctrl/handlers.py. Colocate
it with its single caller to slim utils/ down to real cross-cutting helpers.

Refs #441
