+++ widgets/panels/mana_rich_text_ctrl/handlers.py
@@ -14,7 +14,7 @@ from utils.constants import (
     MANA_TRAILING_SYMBOL_PATTERN,
     NAVIGATION_KEYS,
 )
-from utils.keyboard_evts import key_char
+from widgets.panels.mana_rich_text_ctrl.keyboard_evts import key_char
 
 if TYPE_CHECKING:
     from utils.mana_icon_factory import ManaIconFactory
