+++ widgets/panels/mana_rich_text_ctrl/keyboard_evts.py
